--- a/osu.Framework/Graphics/Containers/TextFlowContainer.cs
+++ b/osu.Framework/Graphics/Containers/TextFlowContainer.cs
@@ -21,10 +21,15 @@ namespace osu.Framework.Graphics.Containers
     /// </summary>
     public class TextFlowContainer : FillFlowContainer
     {
+        protected override Container<Drawable> Content => content;
+        private readonly Container content;
+
         private float firstLineIndent;
         private readonly Action<SpriteText> defaultCreationParameters;
 
         private readonly List<ITextPart> parts = new List<ITextPart>();
+
+        public override IEnumerable<Drawable> FlowingChildren => AliveChildren.Where(d => d.IsPresent).OrderBy(d => LayoutChildren[d]).ThenBy(d => d.ChildID);
 
         private readonly Cached partsCache = new Cached();
 
@@ -115,6 +120,9 @@ namespace osu.Framework.Graphics.Containers
 
                 textAnchor = value;
 
+                Content.Anchor = value;
+                Content.Origin = value;
+
                 layout.Invalidate();
             }
         }
@@ -141,6 +149,12 @@ namespace osu.Framework.Graphics.Containers
 
         public TextFlowContainer(Action<SpriteText> defaultCreationParameters = null)
         {
+            content = new Container
+            {
+                AutoSizeAxes = Axes.Both,
+            };
+            AddInternal(content);
+
             this.defaultCreationParameters = defaultCreationParameters;
         }
 
@@ -174,33 +188,23 @@ namespace osu.Framework.Graphics.Containers
                 RecreateAllParts();
         }
 
-        public override IEnumerable<Drawable> FlowingChildren
-        {
-            get
-            {
-                if ((TextAnchor & (Anchor.x2 | Anchor.y2)) == 0)
-                    return base.FlowingChildren;
-
-                var childArray = base.FlowingChildren.ToArray();
-
-                if ((TextAnchor & Anchor.x2) > 0)
-                    reverseHorizontal(childArray);
-                if ((TextAnchor & Anchor.y2) > 0)
-                    reverseVertical(childArray);
-
-                return childArray;
-            }
-        }
-
         protected override void UpdateAfterChildren()
         {
+            base.UpdateAfterChildren();
+
             if (!layout.IsValid)
             {
                 computeLayout();
                 layout.Validate();
             }
+        }
 
-            base.UpdateAfterChildren();
+        protected override void PerformLayout()
+        {
+            base.PerformLayout();
+
+            computeLayout();
+            layout.Validate();
         }
 
         protected override int Compare(Drawable x, Drawable y)
@@ -284,8 +288,19 @@ namespace osu.Framework.Graphics.Containers
             throw new InvalidOperationException($"Use {nameof(AddText)} to add text to a {nameof(TextFlowContainer)}.");
         }
 
+        public override bool Remove(Drawable drawable, bool disposeImmediately)
+        {
+            LayoutChildren.Remove(drawable);
+            InvalidateLayout();
+
+            return base.Remove(drawable, disposeImmediately);
+        }
+
         public override void Clear(bool disposeChildren)
         {
+            LayoutChildren.Clear();
+            InvalidateLayout();
+
             base.Clear(disposeChildren);
             parts.Clear();
         }
@@ -337,33 +352,12 @@ namespace osu.Framework.Graphics.Containers
         {
             part.RecreateDrawablesFor(this);
             foreach (var drawable in part.Drawables)
-                base.Add(drawable);
-        }
-
-        private void reverseHorizontal(Drawable[] children)
-        {
-            int reverseStartIndex = 0;
-
-            // Inverse the order of all children when displaying backwards, stopping at newline boundaries
-            for (int i = 0; i < children.Length; i++)
             {
-                if (!(children[i] is NewLineContainer))
-                    continue;
-
-                Array.Reverse(children, reverseStartIndex, i - reverseStartIndex);
-                reverseStartIndex = i + 1;
+                base.Add(drawable);
+                LayoutChildren.Add(drawable, 0f);
             }
 
-            // Extra loop for the last newline boundary (or all children if there are no newlines)
-            Array.Reverse(children, reverseStartIndex, children.Length - reverseStartIndex);
-        }
-
-        private void reverseVertical(Drawable[] children)
-        {
-            // A vertical reverse reverses the order of the newline sections, but not the order within the newline sections
-            // For code clarity this is done by reversing the entire array, and then reversing within the newline sections to restore horizontal order
-            Array.Reverse(children);
-            reverseHorizontal(children);
+            InvalidateLayout();
         }
 
         private readonly Cached layout = new Cached();
@@ -375,9 +369,6 @@ namespace osu.Framework.Graphics.Containers
 
             foreach (var c in Children)
             {
-                c.Anchor = TextAnchor;
-                c.Origin = TextAnchor;
-
                 if (c is NewLineContainer nlc)
                 {
                     curLine.Add(nlc);
@@ -411,6 +402,10 @@ namespace osu.Framework.Graphics.Containers
                 float currentLineHeight = 0f;
                 float lineSpacingValue = lastLineHeight * LineSpacing;
 
+                // Compute the offset of this line from the right
+                Drawable lastTextPartInLine = (line[^1] is NewLineContainer && line.Count >= 2) ? line[^2] : line[^1];
+                float lineOffsetFromRight = Content.ChildSize.X - (lastTextPartInLine.X + lastTextPartInLine.DrawWidth);
+
                 foreach (Drawable c in line)
                 {
                     if (c is NewLineContainer nlc)
@@ -430,6 +425,11 @@ namespace osu.Framework.Graphics.Containers
 
                     if (c.Height > currentLineHeight)
                         currentLineHeight = c.Height;
+
+                    if ((TextAnchor & Anchor.x1) != 0)
+                        c.X += lineOffsetFromRight / 2;
+                    else if ((TextAnchor & Anchor.x2) != 0)
+                        c.X += lineOffsetFromRight;
 
                     isFirstChild = false;
                 }