- Closes #5084
- Closes #4753 
- Closes #5499 

The idea here is to put `TextFlowContainer`'s Children inside a container and set its origin and anchor to 'TextAnchor' while the `Children` are all anchored to the top left, this way when `FillFlowContainer.ComputeLayoutPositions()` is called the text will always flow from left to right and from top to bottom.