@bdach 2022-11-15
Benchmark pushed in [a7d0765](https://github.com/ppy/osu-framework/pull/5507/commits/a7d0765c303333e2cafe652dda92e716afdc5571).

Current master:

|  Method |     Mean |   Error |   StdDev | Allocated |
|-------- |---------:|--------:|---------:|----------:|
| SetText | 116.8 ns | 9.12 ns | 26.60 ns |     153 B |

This PR:

|  Method |     Mean |    Error |   StdDev |   Median | Allocated |
|-------- |---------:|---------:|---------:|---------:|----------:|
| SetText | 172.6 ns | 23.98 ns | 68.79 ns | 139.3 ns |     153 B |

Looks like there is a fair bit of overhead but again, likely not a huge deal given invalidations.

Gonna request a second review to ensure I haven't screwed the testing methodology, but looks fine as far as I can tell.